requirements.txt
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.1.0
attrs==22.1.0
Automat==22.10.0
autopep8==2.0.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.3
cssselect==1.2.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.3
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-filter==22.1
django-js-asset==2.0.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
entrypoints==0.4
executing==1.2.0
fake-useragent==0.1.14
fastjsonschema==2.16.2
filelock==3.8.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==22.10.0
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itemadapter==0.7.0
itemloaders==1.0.6
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
packaging==21.3
pandocfilters==1.5.0
parse==1.19.0
parsel==1.7.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
prometheus-client==0.15.0
prompt-toolkit==3.0.33
Protego==0.2.1
psutil==5.9.4
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
PyDispatcher==2.0.6
pyee==8.2.2
Pygments==2.13.0
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
requests-html==0.10.0
rfc3986==1.5.0
Scrapy==2.7.1
scrapy-djangoitem==1.1.1
scrapy-splash==0.8.0
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.1
terminado==0.17.0
tinycss2==1.2.1
tldextract==3.4.0
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
Twisted==22.10.0
typing_extensions==4.4.0
urllib3==1.26.12
w3lib==2.0.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
whitenoise==6.2.0
widgetsnbextension==4.0.3
word2number==1.1
zipp==3.10.0
zope.interface==5.5.1
